Include cluster.m as per the existing mex/linkagemex.cpp
This completes the clustering functionality that OpenTLD needs.

Note that this is a standard matlab .m file so unlike linkagemex.cpp its therefore not optimized. However this may not such be a concern as linkagemex is probably doing the most work by first building the cluster tree. The missing cluster() function then splits the cluster built be linkagemex.

See:

http://www.mathworks.com/help/toolbox/stats/linkage.html
http://www.mathworks.com/help/toolbox/stats/cluster.html